move download to vision
